requirements.txt
channels
channels-redis
daphne
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
gunicorn
PyJWT
redis